Creados tests para la parte implementada de mesas Issue #21

+++ Proyecto/src/Mesas/mesa.java
@@ -22,7 +22,7 @@ public class mesa {
 		return this.pedido;
 	}
 	
-	public void aniadirPlato(String nombre,float precio,int cantidad) throws JSONException {
+	public void aniadirPlato(String nombre,double precio,int cantidad) throws JSONException {
 		JSONObject nuevoped = new JSONObject();
 		nuevoped.put("Plato", nombre);
 		nuevoped.put("Precio", precio);
